Add files via upload
Analyze A/B Test Results: 
A/B tests are very commonly performed by data analysts and data scientists. I
In this project I tried to understand the results of an A/B test run by an e-commerce website. The company has developed a new web page in order to try and increase the number of users who "convert," meaning the number of users who decide to pay for the company's product. My goal was to work through this notebook to help the company understand if they should implement this new page, keep the old page, or perhaps run the experiment longer to make their decision.